Move RedisCallback from rdb_reader to redis module
This mirrors the location of MemoryCallback and highlights that this
is a backend-specific class.
